code organization
creating a System, classStack classes
adding a commandRegistry object
in seperate files
